Add ability to insert a new journal Entry
Why
---
I've implmented the "add entry" command using the openEditorIfNotAlreadyOpen
function so that creating an entry is naturally led into creating a journal
first if it doesnt exist. The intent of the 4-5 lines at the end of newEntry is
to create a new line at the end and attach a timestamp to it.